% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_Piper.R
\name{ggplot_piper}
\alias{ggplot_piper}
\title{create piper plot}
\usage{
ggplot_piper()
}
\value{
empty piper plot
}
\description{
create piper plot
}
\examples{
# empty piper plot
ggplot_piper()
# basic pot with values
ggplot_piper() + geom_point(aes(x,y), data=piper_data)
# This plot behaves like a regular ggplot. You can customize it according to your preferences 
}
