% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_Piper.R
\name{transform_piper_data}
\alias{transform_piper_data}
\title{transform_piper_data}
\usage{
transform_piper_data(Mg, Ca, Cl, SO4, name = NULL)
}
\arguments{
\item{Mg}{}

\item{Ca}{}

\item{Cl}{}

\item{SO4}{}

\item{name}{}
}
\value{
dataframe with coordinates
}
\description{
transform_piper_data
}
\examples{
transform_piper_data(Ca   = data$Ca, Mg   = data$Mg, Cl = data$Cl, SO4  = data$SO4, name = data$IDs)
}
